requirements.txt
python-decouple>=3.8, <3.9
uvicorn>=0.22.0, <0.23
fastapi>=0.97, <1.0
python-multipart>=0.0.6, <0.1
requests>=2.31.0, <2.32
httpx>=0.24.1, <0.25
asgi-correlation-id>=4.2.0, <4.3
celery>=5.3.0, <5.4
dask[dataframe]>=2023.6.0, <2024
black
ruff
pytest>=7.3.2, <7.4
